feat: add the logic of level 1 task can have point initially but can't if it does have children.
